+++ src/action.spec.ts
@@ -6,11 +6,14 @@ import { AbstractLogger } from "./logger";
 describe("Action", () => {
   const simpleHandler = vi.fn();
   const simpleAction = new Action({
+    event: "simple",
     input: z.tuple([z.string()]),
     handler: simpleHandler,
   });
   const ackHandler = vi.fn(async (): Promise<[number]> => [123]);
   const ackAction = new Action({
+    event: "ackOne",
+    ns: "test",
     input: z.tuple([z.string()]),
     output: z.tuple([z.number()]),
     handler: ackHandler,
@@ -20,6 +23,20 @@ describe("Action", () => {
     test("should create inheritor of AbstractAction", () => {
       expect(simpleAction).toBeInstanceOf(Action);
       expect(simpleAction).toBeInstanceOf(AbstractAction);
+    });
+  });
+
+  describe("getName()", () => {
+    test("should return the event name", () => {
+      expect(simpleAction.getEvent()).toBe("simple");
+      expect(ackAction.getEvent()).toBe("ackOne");
+    });
+  });
+
+  describe("getNamespace()", () => {
+    test("should return the namespace", () => {
+      expect(simpleAction.getNamespace()).toBe("/");
+      expect(ackAction.getNamespace()).toBe("test");
     });
   });