+++ src/actions-factory.ts
@@ -3,35 +3,54 @@ import { Action } from "./action";
 import { Config } from "./config";
 import { EmissionMap } from "./emission";
 import { ActionContext, Handler } from "./handler";
+import { Namespaces, RootNS } from "./namespaces";
+
+interface Commons<
+  IN extends z.AnyZodTuple,
+  NS extends Namespaces<EmissionMap>,
+  K extends keyof NS,
+> {
+  /** @desc The incoming event payload validation schema (without or excluding acknowledgement) */
+  input: IN;
+  /**
+   * @desc The namespace this Action belongs to (optional)
+   * @default "/"
+   * */
+  ns?: K;
+  /** @desc The incoming event name */
+  event: string;
+}
 
 export interface ActionNoAckDef<
   IN extends z.AnyZodTuple,
-  E extends EmissionMap,
-> {
-  /** @desc The incoming event payload validation schema (no acknowledgement) */
-  input: IN;
+  NS extends Namespaces<EmissionMap>,
+  K extends keyof NS,
+> extends Commons<IN, NS, K> {
   /** @desc No output schema => no returns => no acknowledgement */
-  handler: Handler<ActionContext<z.output<IN>, E>, void>;
+  handler: Handler<ActionContext<z.output<IN>, NS[K]>, void>;
 }
 
 export interface ActionWithAckDef<
   IN extends z.AnyZodTuple,
   OUT extends z.AnyZodTuple,
-  E extends EmissionMap,
-> {
-  /** @desc The incoming event payload (excl. acknowledgement) validation schema */
-  input: IN;
+  NS extends Namespaces<EmissionMap>,
+  K extends keyof NS,
+> extends Commons<IN, NS, K> {
   /** @desc The acknowledgement validation schema */
   output: OUT;
   /** @desc The returns become an Acknowledgement */
-  handler: Handler<ActionContext<z.output<IN>, E>, z.input<OUT>>;
+  handler: Handler<ActionContext<z.output<IN>, NS[K]>, z.input<OUT>>;
 }
 
-export class ActionsFactory<E extends EmissionMap> {
-  constructor(protected config: Config<E>) {}
+export class ActionsFactory<NS extends Namespaces<EmissionMap>> {
+  constructor(protected config: Config<NS>) {}
 
-  public build<IN extends z.AnyZodTuple, OUT extends z.AnyZodTuple>(
-    def: ActionNoAckDef<IN, E> | ActionWithAckDef<IN, OUT, E>,
+  public build<
+    IN extends z.AnyZodTuple,
+    OUT extends z.AnyZodTuple,
+    K extends keyof NS & string = RootNS,
+  >(
+    def: ActionNoAckDef<IN, NS, K> | ActionWithAckDef<IN, OUT, NS, K>,
   ): Action<IN, OUT> {
     return new Action(def);
   }