+++ src/config.spec.ts
@@ -1,4 +1,5 @@
 import { describe, expect, test, vi } from "vitest";
+import { z } from "zod";
 import { createConfig } from "./config";
 import { AbstractLogger } from "./logger";
 
@@ -7,12 +8,30 @@ describe("Config", () => {
     test("should return its argument", () => {
       expect(
         createConfig({
-          emission: {},
+          emission: {
+            "/": {},
+            test: {},
+          },
           timeout: 2000,
           logger: { debug: vi.fn() } as unknown as AbstractLogger,
         }),
       ).toEqual({
-        emission: {},
+        emission: { "/": {}, test: {} },
+        timeout: 2000,
+        logger: { debug: expect.any(Function) },
+      });
+    });
+
+    test("should ensure namespaces", () => {
+      const schema = z.tuple([]);
+      expect(
+        createConfig({
+          emission: { test: { schema } },
+          timeout: 2000,
+          logger: { debug: vi.fn() } as unknown as AbstractLogger,
+        }),
+      ).toEqual({
+        emission: { "/": { test: { schema } } },
         timeout: 2000,
         logger: { debug: expect.any(Function) },
       });